pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]==7.1.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mx_bluesky"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
description = "Bluesky tools for MX Beamlines"
dependencies = [
    "bluesky",
    "ophyd",
    "pyepics",
    "caproto",
    "jupyterlab",
    "matplotlib",
    "requests",
    "opencv-python",
    "pydantic",
    "dls-dodal @ git+https://github.com/DiamondLightSource/dodal.git@main",
    "blueapi",
]
dynamic = ["version"]
license.file = "LICENSE"
readme = "README.rst"
requires-python = ">=3.10"

[project.optional-dependencies]
dev = [
    "GitPython",
    "black",
    "mypy",
    "pipdeptree",
    "pre-commit",
    "pydata-sphinx-theme>=0.12",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "sphinx-design",
    "tox-direct",
    "types-mock",
    "types-requests",
]

[project.scripts]
mx_bluesky = "mx_bluesky.__main__:main"
run_extruder = "mx_bluesky.I24.serial.run_serial:run_extruder"
run_fixed_target = "mx_bluesky.I24.serial.run_serial:run_fixed_target"

[project.urls]
GitHub = "https://github.com/DiamondLightSource/mx_bluesky"

[[project.authors]] # Further authors may be added by duplicating this section
email = "dominic.oram@diamond.ac.uk"
name = "Dominic Oram"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/mx_bluesky/_version.py"

[tool.mypy]
ignore_missing_imports = true # Ignore missing stubs in imported modules

[tool.ruff]
src = ["src", "tests"]
line-length = 88
lint.extend-ignore = [
    "E501", # Line too long
    "F811", # support typing.overload decorator
]
lint.select = [
    "C4",   # flake8-comprehensions - https://beta.ruff.rs/docs/rules/#flake8-comprehensions-c4
    "E",    # pycodestyle errors - https://beta.ruff.rs/docs/rules/#error-e
    "F",    # pyflakes rules - https://beta.ruff.rs/docs/rules/#pyflakes-f
    "W",    # pycodestyle warnings - https://beta.ruff.rs/docs/rules/#warning-w
    "I001", # isort
]

[tool.pytest.ini_options]
# Run pytest with all our checkers, and don't spam us with massive tracebacks on error
asyncio_mode = "auto"
addopts = """
    --tb=native -vv --doctest-modules --doctest-glob="*.rst"
    --cov=mx_bluesky --cov-report term --cov-report xml:cov.xml
    """
# https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
#filterwarnings = "error"
# Doctest python code in docs, python code in src docstrings, test functions in tests
testpaths = "docs src tests"

[tool.coverage.run]
data_file = "/tmp/mx_bluesky.coverage"

[tool.coverage.paths]
# Tests are run from installed location, map back to the src directory
source = ["src", "**/site-packages/"]

# tox must currently be configured via an embedded ini string
# See: https://github.com/tox-dev/tox/issues/999
[tool.tox]
legacy_tox_ini = """
[tox]
skipsdist=True

[testenv:{pre-commit,mypy,pytest,docs}]
# Don't create a virtualenv for the command, requires tox-direct plugin
direct = True
passenv = *
allowlist_externals =
    pytest
    pre-commit
    mypy
    sphinx-build
    sphinx-autobuild
commands =
    pytest: pytest {posargs}
    mypy: mypy src tests {posargs} --ignore-missing-imports --no-strict-optional {posargs}
    pre-commit: pre-commit run --all-files {posargs}
    docs: sphinx-{posargs:build -EW --keep-going} -T docs build/html
"""
